Created Separate Activities for pictures, contacts, and messages; Added template layouts for each one.
