+++ app/src/main/java/edu/njit/cs656/nfcapp/SendContactActivity.java
@@ -1,0 +1,27 @@
+package edu.njit.cs656.nfcapp;
+
+import android.content.Intent;
+import android.net.Uri;
+import android.nfc.NfcAdapter;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.EditText;
+
+import static android.provider.AlarmClock.EXTRA_MESSAGE;
+
+public class SendContactActivity extends AppCompatActivity {
+
+    private Uri[] mFileUris = new Uri[10];
+    private NfcAdapter mNfcAdapter;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_send_contact);
+    }
+
+//    public void send(View view) {
+//    }
+
+}
