+++ app/src/main/res/layout/activity_send_picture.xml
@@ -1,0 +1,33 @@
+ <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:orientation="vertical"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
+        >
+
+     <TextView android:id="@+id/textview"
+         android:layout_width="fill_parent"
+         android:layout_height="512dp"
+         android:text="This should be a scrollable window that displays the pictures - currently hardcoded to 1 picture"
+         android:textSize="18dp"
+         android:gravity="center"
+         android:layout_marginTop="10dp"
+         />
+
+     <Button
+         android:id="@+id/button3"
+         android:layout_width="match_parent"
+         android:layout_height="64dp"
+         android:onClick="send"
+         android:padding="8dp"
+         android:text="Send"
+         android:textSize="24dp" />
+
+     <RelativeLayout android:id="@+id/relativeLayout1"
+            android:layout_height="wrap_content"
+            android:layout_width="wrap_content">
+
+     </RelativeLayout>
+
+
+ </LinearLayout>
+
